+++ engine/world/named_entity.py
@@ -1,0 +1,4 @@
+class NamedEntity:
+    def __init__(self, id: str, name: str):
+        self.id = id
+        self.name = name
