+++ night_of_1000_dans/prologue/title.py
@@ -1,0 +1,46 @@
+from engine.scene import Scene
+from engine.printer import print, input
+from night_of_1000_dans.prologue.create_character import CreateCharacterScene
+
+
+class TitleScene(Scene):
+    def run(self) -> CreateCharacterScene:
+        print(
+            """
+ ███▄    █  ██▓  ▄████  ██░ ██ ▄▄▄█████▓    ▒█████    █████▒    ▄▄▄
+ ██ ▀█   █ ▓██▒ ██▒ ▀█▒▓██░ ██▒▓  ██▒ ▓▒   ▒██▒  ██▒▓██   ▒    ▒████▄
+▓██  ▀█ ██▒▒██▒▒██░▄▄▄░▒██▀▀██░▒ ▓██░ ▒░   ▒██░  ██▒▒████ ░    ▒██  ▀█▄
+▓██▒  ▐▌██▒░██░░▓█  ██▓░▓█ ░██ ░ ▓██▓ ░    ▒██   ██░░▓█▒  ░    ░██▄▄▄▄██
+▒██░   ▓██░░██░░▒▓███▀▒░▓█▒░██▓  ▒██▒ ░    ░ ████▓▒░░▒█░        ▓█   ▓██▒
+░ ▒░   ▒ ▒ ░▓   ░▒   ▒  ▒ ░░▒░▒  ▒ ░░      ░ ▒░▒░▒░  ▒ ░        ▒▒   ▓▒█░
+░ ░░   ░ ▒░ ▒ ░  ░   ░  ▒ ░▒░ ░    ░         ░ ▒ ▒░  ░           ▒   ▒▒ ░
+   ░   ░ ░  ▒ ░░ ░   ░  ░  ░░ ░  ░         ░ ░ ░ ▒   ░ ░         ░   ▒
+         ░  ░        ░  ░  ░  ░                ░ ░                   ░  ░
+
+▄▄▄█████▓ ██░ ██  ▒█████   █    ██   ██████  ▄▄▄       ███▄    █ ▓█████▄
+▓  ██▒ ▓▒▓██░ ██▒▒██▒  ██▒ ██  ▓██▒▒██    ▒ ▒████▄     ██ ▀█   █ ▒██▀ ██▌
+▒ ▓██░ ▒░▒██▀▀██░▒██░  ██▒▓██  ▒██░░ ▓██▄   ▒██  ▀█▄  ▓██  ▀█ ██▒░██   █▌
+░ ▓██▓ ░ ░▓█ ░██ ▒██   ██░▓▓█  ░██░  ▒   ██▒░██▄▄▄▄██ ▓██▒  ▐▌██▒░▓█▄   ▌
+  ▒██▒ ░ ░▓█▒░██▓░ ████▓▒░▒▒█████▓ ▒██████▒▒ ▓█   ▓██▒▒██░   ▓██░░▒████▓
+  ▒ ░░    ▒ ░░▒░▒░ ▒░▒░▒░ ░▒▓▒ ▒ ▒ ▒ ▒▓▒ ▒ ░ ▒▒   ▓▒█░░ ▒░   ▒ ▒  ▒▒▓  ▒
+    ░     ▒ ░▒░ ░  ░ ▒ ▒░ ░░▒░ ░ ░ ░ ░▒  ░ ░  ▒   ▒▒ ░░ ░░   ░ ▒░ ░ ▒  ▒
+  ░       ░  ░░ ░░ ░ ░ ▒   ░░░ ░ ░ ░  ░  ░    ░   ▒      ░   ░ ░  ░ ░  ░
+          ░  ░  ░    ░ ░     ░           ░        ░  ░         ░    ░
+                                                                  ░
+▓█████▄  ▄▄▄       ███▄    █   ██████
+▒██▀ ██▌▒████▄     ██ ▀█   █ ▒██    ▒
+░██   █▌▒██  ▀█▄  ▓██  ▀█ ██▒░ ▓██▄
+░▓█▄   ▌░██▄▄▄▄██ ▓██▒  ▐▌██▒  ▒   ██▒
+░▒████▓  ▓█   ▓██▒▒██░   ▓██░▒██████▒▒
+ ▒▒▓  ▒  ▒▒   ▓▒█░░ ▒░   ▒ ▒ ▒ ▒▓▒ ▒ ░
+ ░ ▒  ▒   ▒   ▒▒ ░░ ░░   ░ ▒░░ ░▒  ░ ░
+ ░ ░  ░   ░   ▒      ░   ░ ░ ░  ░  ░
+   ░          ░  ░         ░       ░
+ ░
+        """,
+            delay=0.001,
+        )
+        print("Off-By-One Entertainment.  All Rights Reserved.\n")
+        print("Press enter to continue...")
+        input(None)
+        return CreateCharacterScene()
