+++ night_of_1000_dans/world/coffee/scenes/coffee.py
@@ -1,0 +1,12 @@
+from engine.scene import Scene
+from engine.printer import print
+from night_of_1000_dans.state import game_state
+
+
+class CoffeeScene(Scene):
+    def run(self):
+        print(
+            "You run next door to Gotham Coffee and promptly hand over no less than $7.15 for a cup of joe."
+        )
+        print("Worth it.")
+        return game_state.previous_scene
