--- persons/model.py
@@ -1,12 +1,0 @@
-class Person:
-    def __init__(self, id: int, name: str):
-        self.id = id
-        self.name = name
-
-
-joy = Person(897123, "Joy")
-dee = Person(12432132, "Dee")
-dan_e = Person(12351235, "Daniel E")
-dan_b = Person(1253235, "Dan B")
-dan_i = Person(1253, "Dan I")
-dan_s = Person(184726, "Dan S")
