--- scenes/title.py
@@ -1,47 +1,0 @@
-from printer import input, print
-
-from scenes.create_character import CreateCharacterScene
-from scenes.model import Scene
-
-
-class TitleScene(Scene):
-    def run(self) -> CreateCharacterScene:
-        print(
-            """
- ███▄    █  ██▓  ▄████  ██░ ██ ▄▄▄█████▓    ▒█████    █████▒    ▄▄▄
- ██ ▀█   █ ▓██▒ ██▒ ▀█▒▓██░ ██▒▓  ██▒ ▓▒   ▒██▒  ██▒▓██   ▒    ▒████▄
-▓██  ▀█ ██▒▒██▒▒██░▄▄▄░▒██▀▀██░▒ ▓██░ ▒░   ▒██░  ██▒▒████ ░    ▒██  ▀█▄
-▓██▒  ▐▌██▒░██░░▓█  ██▓░▓█ ░██ ░ ▓██▓ ░    ▒██   ██░░▓█▒  ░    ░██▄▄▄▄██
-▒██░   ▓██░░██░░▒▓███▀▒░▓█▒░██▓  ▒██▒ ░    ░ ████▓▒░░▒█░        ▓█   ▓██▒
-░ ▒░   ▒ ▒ ░▓   ░▒   ▒  ▒ ░░▒░▒  ▒ ░░      ░ ▒░▒░▒░  ▒ ░        ▒▒   ▓▒█░
-░ ░░   ░ ▒░ ▒ ░  ░   ░  ▒ ░▒░ ░    ░         ░ ▒ ▒░  ░           ▒   ▒▒ ░
-   ░   ░ ░  ▒ ░░ ░   ░  ░  ░░ ░  ░         ░ ░ ░ ▒   ░ ░         ░   ▒
-         ░  ░        ░  ░  ░  ░                ░ ░                   ░  ░
-
-▄▄▄█████▓ ██░ ██  ▒█████   █    ██   ██████  ▄▄▄       ███▄    █ ▓█████▄
-▓  ██▒ ▓▒▓██░ ██▒▒██▒  ██▒ ██  ▓██▒▒██    ▒ ▒████▄     ██ ▀█   █ ▒██▀ ██▌
-▒ ▓██░ ▒░▒██▀▀██░▒██░  ██▒▓██  ▒██░░ ▓██▄   ▒██  ▀█▄  ▓██  ▀█ ██▒░██   █▌
-░ ▓██▓ ░ ░▓█ ░██ ▒██   ██░▓▓█  ░██░  ▒   ██▒░██▄▄▄▄██ ▓██▒  ▐▌██▒░▓█▄   ▌
-  ▒██▒ ░ ░▓█▒░██▓░ ████▓▒░▒▒█████▓ ▒██████▒▒ ▓█   ▓██▒▒██░   ▓██░░▒████▓
-  ▒ ░░    ▒ ░░▒░▒░ ▒░▒░▒░ ░▒▓▒ ▒ ▒ ▒ ▒▓▒ ▒ ░ ▒▒   ▓▒█░░ ▒░   ▒ ▒  ▒▒▓  ▒
-    ░     ▒ ░▒░ ░  ░ ▒ ▒░ ░░▒░ ░ ░ ░ ░▒  ░ ░  ▒   ▒▒ ░░ ░░   ░ ▒░ ░ ▒  ▒
-  ░       ░  ░░ ░░ ░ ░ ▒   ░░░ ░ ░ ░  ░  ░    ░   ▒      ░   ░ ░  ░ ░  ░
-          ░  ░  ░    ░ ░     ░           ░        ░  ░         ░    ░
-                                                                  ░
-▓█████▄  ▄▄▄       ███▄    █   ██████
-▒██▀ ██▌▒████▄     ██ ▀█   █ ▒██    ▒
-░██   █▌▒██  ▀█▄  ▓██  ▀█ ██▒░ ▓██▄
-░▓█▄   ▌░██▄▄▄▄██ ▓██▒  ▐▌██▒  ▒   ██▒
-░▒████▓  ▓█   ▓██▒▒██░   ▓██░▒██████▒▒
- ▒▒▓  ▒  ▒▒   ▓▒█░░ ▒░   ▒ ▒ ▒ ▒▓▒ ▒ ░
- ░ ▒  ▒   ▒   ▒▒ ░░ ░░   ░ ▒░░ ░▒  ░ ░
- ░ ░  ░   ░   ▒      ░   ░ ░ ░  ░  ░
-   ░          ░  ░         ░       ░
- ░
-        """,
-            delay=0.001,
-        )
-        print("Off-By-One Entertainment.  All Rights Reserved.\n")
-        print("Press enter to continue...")
-        input(None)
-        return CreateCharacterScene()
